api/requirements.txt
fastapi>=0.121.2
uvicorn[standard]>=0.38.0
requests>=2.32.5
python-dotenv>=1.2.1
google-genai>=1.50.1
